+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/connect/analytics/_components/ConnectAnalyticsDashboard.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { mobileViewport } from "../../../../../../../stories/utils";
-import { ConnectAnalyticsDashboard } from "../ConnectAnalyticsDashboard";
-import { createWalletStatusStub } from "./storyUtils";
+import { ConnectAnalyticsDashboardUI } from "../ConnectAnalyticsDashboardUI";
+import { createWalletStatsStub } from "./storyUtils";
 
 const meta = {
   title: "Charts/Connect/Analytics Dashboard",
@@ -28,8 +28,8 @@ export const Mobile: Story = {
 function Component() {
   return (
     <div className="min-h-screen bg-background p-4 text-foreground gap-6 max-w-[1000px] flex flex-col mx-auto py-10">
-      <ConnectAnalyticsDashboard
-        walletStats={createWalletStatusStub(30)}
+      <ConnectAnalyticsDashboardUI
+        walletStats={createWalletStatsStub(30)}
         isLoading={false}
       />
     </div>
