+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/connect/analytics/_components/DailyConnectionsChartCard.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   type ChartConfig,
   ChartContainer,
