+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/connect/analytics/_components/WalletConnectorsChartChart.stories.tsx
@@ -4,7 +4,7 @@ import {
   mobileViewport,
 } from "../../../../../../../stories/utils";
 import { WalletConnectorsChartCard } from "./WalletConnectorsChartCard";
-import { createWalletStatusStub } from "./storyUtils";
+import { createWalletStatsStub } from "./storyUtils";
 
 const meta = {
   title: "Charts/Connect/Wallet Connectors",
@@ -33,35 +33,35 @@ function Component() {
     <div className="min-h-screen bg-background p-4 text-foreground gap-6 max-w-[1000px] flex flex-col mx-auto">
       <BadgeContainer label="30 days">
         <WalletConnectorsChartCard
-          walletStats={createWalletStatusStub(30)}
+          walletStats={createWalletStatsStub(30)}
           isLoading={false}
         />
       </BadgeContainer>
 
       <BadgeContainer label="60 days">
         <WalletConnectorsChartCard
-          walletStats={createWalletStatusStub(60)}
+          walletStats={createWalletStatsStub(60)}
           isLoading={false}
         />
       </BadgeContainer>
 
       <BadgeContainer label="10 days">
         <WalletConnectorsChartCard
-          walletStats={createWalletStatusStub(10)}
+          walletStats={createWalletStatsStub(10)}
           isLoading={false}
         />
       </BadgeContainer>
 
       <BadgeContainer label="0 days">
         <WalletConnectorsChartCard
-          walletStats={createWalletStatusStub(0)}
+          walletStats={createWalletStatsStub(0)}
           isLoading={false}
         />
       </BadgeContainer>
 
       <BadgeContainer label="Loading">
         <WalletConnectorsChartCard
-          walletStats={createWalletStatusStub(0)}
+          walletStats={createWalletStatsStub(0)}
           isLoading={true}
         />
       </BadgeContainer>
