+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/connect/layout.tsx
@@ -12,12 +12,12 @@ export default function Layout(props: {
 
   const links: { label: string; href: string }[] = [
     {
-      label: "In-App Wallets",
-      href: `/team/${team_slug}/${project_slug}/connect/in-app-wallets`,
-    },
-    {
       label: "Analytics",
       href: `/team/${team_slug}/${project_slug}/connect/analytics`,
+    },
+    {
+      label: "In-App Wallets",
+      href: `/team/${team_slug}/${project_slug}/connect/in-app-wallets`,
     },
     {
       label: "Ecosystem Wallets",
